lib/syscall_shim: uk_prsyscall: Fix typecast for PT_BUFP
PT_BUFP prints binary buffer content human readable but had one
particular problem: Due to incomplete typecasting, printing of a
negative byte (__s8) caused printing a 4-byte hex sequence instead
of just one byte with the format: `\0xHH`.
This commit fixes this with proper typecasting.

Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #750

+++ lib/syscall_shim/uk_prsyscall.c
@@ -624,7 +624,7 @@ static void pr_param(struct uk_streambuf *sb, int fmtf,
 					uk_streambuf_printf(sb, "%c", *c);
 				else
 					uk_streambuf_printf(sb, "\\x%02X",
-							    (int) *c);
+							    (int) ((__u8) *c));
 			}
 			uk_streambuf_strcpy(sb, "\"");
 			uk_streambuf_shcc(sb, fmtf, RESET);
